% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edge_corr_wt}
\alias{edge_corr_wt}
\title{Performs the \emph{walkthrough} edge-correction method for a set of sample
trees and a given boundary}
\usage{
edge_corr_wt(dt_s_tree, dt_s_loc, boundary, on = "id_plot")
}
\arguments{
\item{dt_s_tree}{A data.table object as coming from \code{\link{sample_data}}
function}

\item{dt_s_loc}{A data.table object as coming from \code{\link{xy_sample}}
function}

\item{boundary}{A \code{\link[sp]{SpatialPolygons}} object containing one
polygon, describing the boundary of the population.}

\item{on}{Indicate on which column \code{dt_s_tree} and \code{dt_s_loc}
should be joined, see \code{\link[data.table]{data.table}}.}
}
\value{
A vector with row indices identifying trees in dt_s_tree that should
  be counted twice.
}
\description{
Performs the \emph{walkthrough} edge-correction method for a set of sample
trees and a given boundary
}
\references{
Ducey, M.J., Gove, J.H., Valentine, H.T., 2004. A walkthrough
  solution to the boundary overlap problem. Forest Science 50, 427–435.
}
