% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{est_density}
\alias{est_density}
\title{Helper function to estimate density of tree locations per unit area using a
kernel smoothed intensity function}
\usage{
est_density(tree_loc)
}
\arguments{
\item{tree_loc}{A matrix where each row is an individual observation and the
first column is the x-coordinate and the second column is the y-coordinate.}
}
\value{
A summary of the values from pixel image of intensity values.
}
\description{
Helper function to estimate density of tree locations per unit area using a
kernel smoothed intensity function
}
\details{
An object of class \code{\link[spatstat]{ppp}} is created from the
  coordinates in \code{tree_loc} and an pixel image of intensity values is
  created using \code{\link[spatstat]{density.ppp}}.
}
\seealso{
\link[spatstat]{ppp}, \link[spatstat]{density.ppp}
}
