% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_design.R
\name{est_srs}
\alias{est_srs}
\title{Simple random sampling estimator of population means}
\usage{
est_srs(data)
}
\arguments{
\item{data}{A data.table object with plot-level information of stand
characteristics as comin from the \code{plot_data} function.}
}
\value{
A data.table object with mean and variance estimates of variables of
  interest.
}
\description{
Simple random sampling estimator of population means
}
