% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_area}
\alias{extract_area}
\title{Helper function to determine the area of the plot given by the boundary element}
\usage{
extract_area(sf_poly)
}
\arguments{
\item{sf_poly}{An object of class \code{\link[sf]{sf}} and geometry type POLYGON}
}
\value{
A numeric vector with the areas of all polygons in the object.
  Holes within the polygons are already subtracted.
}
\description{
Helper function to determine the area of the plot given by the boundary element
}
\examples{
a <- extract_area(hberg_beech$boundary)

}
