% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_data}
\alias{extract_data}
\title{Helper function to extract population elements into a sample using a list of
indices}
\usage{
extract_data(tree_pop, response)
}
\arguments{
\item{tree_pop}{The population from which to extract sampled elements.}

\item{response}{An integer matrix where rows represent sample locations and columns
represent indices of the trees in \code{dt_pop} that are selected at the
individual sample locations. Zeroes are used to indicate no neighbours and
to ensure a rectangular data format.}
}
\value{
A \code{\link[data.table]{data.table}} object containing the population
  elements from data that were selected into the sample. The data.table
  object has a new variable \code{id_plot} for grouping single sample
  locations.
}
\description{
Helper function to extract population elements into a sample using a list of
indices
}
