% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{is.point_data}
\alias{is.point_data}
\title{Check if an object is a point data object.}
\usage{
is.point_data(x)
}
\arguments{
\item{x}{An \code{R} object}
}
\value{
TRUE if \code{x} is of class \code{\link{point data}};
}
\description{
Check if an object is a point data object.
}
