% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{is.response}
\alias{is.response}
\title{Check if an object is a response design object.}
\usage{
is.response(x)
}
\arguments{
\item{x}{An \code{R} object}
}
\value{
TRUE if \code{x} is of class \code{\link{response}};
}
\description{
Check if an object is a response design object.
}
