% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{is.sample_loc}
\alias{is.sample_loc}
\title{Check if an object is a sample locations object.}
\usage{
is.sample_loc(x)
}
\arguments{
\item{x}{An \code{R} object}
}
\value{
TRUE if \code{x} is of class \code{\link{sample_loc}};
}
\description{
Check if an object is a sample locations object.
}
