% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{is.tree_pop}
\alias{is.tree_pop}
\title{Check if an object is a tree population.}
\usage{
is.tree_pop(x)
}
\arguments{
\item{x}{An \code{R} object}
}
\value{
TRUE if \code{x} is of class \code{\link{tree_pop}};
}
\description{
Check if an object is a tree population.
}
