% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{is.tree_sample}
\alias{is.tree_sample}
\title{Check if an object is a tree sample object.}
\usage{
is.tree_sample(x)
}
\arguments{
\item{x}{An \code{R} object}
}
\value{
TRUE if \code{x} is of class \code{\link{tree_sample}};
}
\description{
Check if an object is a tree sample object.
}
