% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_designs.R
\name{k_tree}
\alias{k_tree}
\title{The \code{k} trees closest to a sample location are selected into the sample}
\usage{
k_tree(tree_loc, sample_loc, k = 6)
}
\arguments{
\item{tree_loc}{A matrix where each is an individual observation and the
first column is the x-coordinate and the second column is the y-coordinate.}

\item{sample_loc}{A matrix with x-coordinates in the first column and
y-coordinates in the second. Each row represents one sample location.}

\item{k}{Number of nearets neighbors.}
}
\value{
An integer matrix where rows represent sample locations and columns
  represent indices of the trees in \code{tree_loc} that are selected at the
  individual sample locations. See the output of the \code{\link[RANN]{nn2}}
  function.
}
\description{
The \code{k} trees closest to a sample location are selected into the sample
}
\details{
The RANN package is used to find the \code{k} trees clostest to the
  respective sample locations in \code{sample_loc}. The RANN package is a
  wrapper for the Approximate Near Neighbor (ANN) C++ library allowing for
  fast nearest-neighbor searches using kd-trees.
}
