% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kalimantan_peat_border}
\alias{kalimantan_peat_border}
\title{The two-dimensional outline of the \code{\link[fisim]{kalimantan_peat}} data
set}
\format{A \code{\link[sp]{SpatialPolygon}} object storing the outline of the
  full-census plot}
\usage{
kalimantan_peat_border
}
\description{
The two-dimensional outline of the \code{\link[fisim]{kalimantan_peat}} data
set
}
\keyword{datasets}
