% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_launcher.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Launch Shiny App}
\usage{
launch_shiny_app()
}
\value{
Launches the Shiny App and returns a shiny application object.
}
\description{
The Shiny App is graphical user interface (GUI) to the functionality of
this package. It plots a map of all trees of the considered stand,
the sample plots and the trees being part of the sample. The user can select
a sampling design and the number of simulations to run. Resulting estimates
are displayed in a table and as a histogram.
}
\examples{
\dontrun{launch_shiny_app()}
}
