% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{point_data}
\alias{point_data}
\title{Create objects of class \code{point_data}}
\usage{
point_data(data, r_design, r_design_parm)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} object holding aggregated
tree data together with identifiers that indicate the sample and the point
location.}

\item{r_design}{A \code{character} string indicating the type of the response
design. One of the following: "fixed_area", "k_tree", "angle_count".}

\item{r_design_parm}{A response design specific parameter of type
\code{numeric}. For "fixed_area" the plot radius in meter; for "k_tree" the
number of trees closest to the sample location; for "angle_count" the basal
area factor.}
}
\value{
\enumerate{
\item \code{$data} A \code{\link[data.table]{data.table}} object as
described above.
\item \code{$r_design}
\item \code{$r_design_parm}
}
}
\description{
The function \code{point_data} creates point data objects. Such objects store
aggregated tree-level information by sample and sample point location. The
function is only used internally by \code{\link{sum_data}}.
}
\details{
The \code{\link[data.table]{data.table}} object (\code{data}) should
  have the following columns:
  \itemize{
  \item \code{id_sample}
  \item \code{id_point}
  \item \code{variable}
  }

  When \code{\link{k_tree}} sampling was applied at the individual sample
  locations, there are four approximations for expanding the target variables
  listed under the \code{variable} column (see \code{\link{sum_k_tree}}).
  The respective column names are:
  \itemize{
  \item \code{Prodan}
  \item \code{Eberhardt}
  \item \code{d_mean}
  \item \code{a_mean}
  }

  For other response designs (e.g., angle count, fixed area), expansion to
  per hectare values is based on known expansion factors that can be derived
  from the response design parameters and not on approximations. In such
  cases, the expanded values are listed under \code{ef_ana}, meaning
  analytically derived expansion factors.

  The last column \code{is_edge} indicates whether some sort of edge
  correction was performed at the respective sample point or not.
}
