% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{response}
\alias{response}
\title{Objects of class response}
\usage{
response(data, r_design, r_design_parm)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} object holding identifiers
for the sample and respective point locations and row indices that indicate
which population elements are selected into the sample.}

\item{r_design}{A \code{character} string indicating the type of the response
design. One of the following: "fixed_area", "k_tree", "angle_count".}

\item{r_design_parm}{A response design specific parameter of type
\code{numeric}. For "fixed_area" the plot radius in meter; for "k_tree" the
number of trees closest to the sample location; for "angle_count" the basal
area factor.}

\item{ef}{The expansion factor, i.e., the factor that is used to prorate tree
attributes to per hectare values (the inverse of the inclusion
probabilities). Provided by the particular response design functions.}

\item{ef_alt1}{Alternative expansion factor that may be applied when using
k-tree sampling.}

\item{ef_alt2}{A third alternative expansion factor approximation for k-tree
sampling}
}
\value{
An object of class \code{response}. A named list with the following
entries:

  \itemize{
  \item \code{$data} A \code{\link[data.table]{data.table}}
  object as described above.
  \item \code{$r_design}
  \item \code{$r_design_parm}
  }
}
\description{
The function \code{response} creates response design objects that store the
results of a particular response design that was applied at given sample
locations to some tree population. The function is only used internally by
the functions that apply the particular response designs:
\code{\link{fixed_area}}, \code{\link{k_tree}}, \code{\link{angle_count}}.
}
\details{
The \code{\link[data.table]{data.table}} object (\code{data}) should
  have the following columns:
  \itemize{
  \item \code{id_sample}
  \item \code{id_point}
  \item \code{s}
  \item \code{ef}
  }

  Some response designs (currently only \code{\link{k_tree}}), may add
  additional alternative expansion factors as optional approximations, where
  unbiased estimators are missing or difficult to derive.
}
