% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{sample_loc}
\alias{sample_loc}
\title{Objects of class sample_loc}
\usage{
sample_loc(data, n, M, method)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} object holding the
cooridnates of sample locations and two identifiers. The first identifier
'id_sample' is used to separate repeated samples of size \code{n} from each
other. There should be \code{M} distinct integers. The second identifier
'id_point' is used to identify individual sample locations in a specific
sample. Both identifiers should be provided as integers. The next two
columns - 'x_s' and 'y_s' - are used for the coordinates and should be
provided as doubles.}

\item{n}{Sample size}

\item{M}{Number of repeated samples}

\item{method}{Were the points spread out randomly or regular}
}
\value{
An object of class \code{sample_loc}. A named list with the following
  four entries:

  \enumerate{
  \item \code{$data} A \code{\link[data.table]{data.table}}
  object containing two identifier columns and two columns for storing
  coordinates
  \item \code{$n} Sample size
  \item \code{$M} Number of repeated
  samples
  \item \code{$method} Random or regular placement of sample
  locations
  }
}
\description{
The function \code{sample_loc} creates sample location objects, named lists
holding coordinates of sample locations spread over some area and
accompanying information about how the samples were generated.
}
\details{
Before object initialization, a couple of checks are performed,
  e.g., column names, data types, data consitency, etc.

  In case random sampling was performed, each sample in \code{data} should
  have exactly \code{n} sample locations. If systematic sampling was applied,
  sample size varies from sample to sample depending on the shape of the
  study region and sample size. If the expected sample size (\code{n})
  provided by the user, deviates by more than 1 from the average sample size
  inferred from \code{data}, a warning is thrown rather than an error.

  The prefered way for creating objects of type sample_loc is via the
  \code{\link{xy_sample}} function.
}
\examples{
library(spatstat);
xy_sample(letterR, n = 10, M = 10, method = "random");
}
