% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_data}
\alias{sum_data}
\title{Helper function to summarize tree-level sample data to the plot-level for
estimation of poplation parameters}
\usage{
sum_data(dt_s_tree, target_vars)
}
\arguments{
\item{dt_s_tree}{A data.table object as coming from \code{\link{sample_data}}
function}

\item{target_var}{A character vector of variable names which should be
summarized at the plot level}
}
\value{
A data.table object with stand characteristics at the plot level.
  Currently, basal area per hectare and the number of trees per hectare are
  included.
}
\description{
Helper function to summarize tree-level sample data to the plot-level for
estimation of poplation parameters
}
