% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_ef}
\alias{sum_ef}
\title{Expansion-factor-based summarizing of tree-level data}
\usage{
sum_ef(tree_sample, target_vars)
}
\arguments{
\item{tree_sample}{A \code{\link{tree_sample}} object as coming from
the \code{\link{extract_data}} function.}

\item{target_vars}{A character vector of variable names which should be
summarized at the plot level}
}
\value{
A data.table object with stand characteristics at the plot level.
  Currently, basal area per hectare and the number of trees per hectare are
  included.
}
\description{
Helper function to summarize tree-level sample data to the plot-level for
estimation of population parameters, where the exact expansion factors are
assumed to be known.
}
