% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_k_tree}
\alias{sum_k_tree}
\title{Summarizing of tree-level data for k-tree sampling}
\usage{
sum_k_tree(tree_sample, target_vars)
}
\arguments{
\item{tree_sample}{A \code{\link{tree_sample}} object as coming from the
\code{\link{extract_data}} function.}

\item{target_vars}{A character vector of variable names which should be
summarized at the plot level}
}
\value{
A data.table object with stand characteristics at the plot level.
  Currently, basal area per hectare and the number of trees per hectare are
  included. The four different approximations are provided under the
  following columns:
  \itemize{
  \item \code{Prodan} - The approach of Prodan, where only half of the k-th
  tree is considered (Prodan, 1968).
  \item \code{Eberhardt} - Expansion is based on the distance to the k-th
  tree and corrected with factor (k-1)/k (Eberhardt 1967)
  \item \code{d_mean} - A circle, where the radius is the average between
  the distances to tree k and tree k+1
  \item \code{a_mean} - The radius of the average area of the circles that
  are built from the distances to tree k and tree k+1
  }
}
\description{
Helper function to summarize tree-level sample data to the plot-level, where
k-tree sampling (\code{\link{k_tree}}) was used for selecting trees at sample
locations.
}
\details{
For k-tree sampling, individual tree inlcusion probabilities can
  only be derived if the tree stem positions of a substantially larger number
  of trees are available than are actually selected at the single sample
  locations (see Kleinn & Vilcko 2006a). In practical timber surveys or
  forest inventories, such extensive numbers of tree positions around the
  actual sample of trees are typically not available. Instead approximations
  for expanding the locally observed forest stand characteristics are
  applied. This functions provides a set of four approximations that are
  presented in Kleinn & Vilcko (2006b).
}
\references{
\itemize{
\item Kleinn, C., Vilčko, F., 2006a. Design-unbiased estimation for
point-to-tree distance sampling. Canadian Journal of Forest Research 36,
1407–1414. https://doi.org/10.1139/x06-038
\item Kleinn, C., Vilčko, F., 2006b. A
new empirical approach for estimation in k-tree sampling. Forest Ecology and
Management 237, 522–533. https://doi.org/10.1016/j.foreco.2006.09.072
\item Prodan, M., 1968. Punktstichprobe für die Forsteinrichtung (A point sample for
forest management planning). Forst Holzwirt 23 (11), 225–226.
\item Eberhardt, L.L., 1967. Some developments in distance sampling. Biometrics 23,
207–216.
}
}
