% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{tree_pop}
\alias{tree_pop}
\title{Objects of class tree_pop}
\usage{
tree_pop(data, boundary)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} object that describes the
individual trees in a forest. Each row represents a tree, while tree
attributes are stored in columns. A minimum set of attributes with specific
columns names is required; see Details section.}

\item{boundary}{A \code{\link[sp]{SpatialPolygons}} object that defines the
border of the forest.}
}
\value{
An object of class \code{tree_pop}. A named list with the following
  two entries:

  \enumerate{
  \item \code{$data} A \code{\link[data.table]{data.table}}
  object as described above.
  \item \code{$boundary} A \code{\link[sp]{SpatialPolygons}} object that
  defines the border of the forest.
  }
}
\description{
The function \code{tree_pop} creates tree population objects holding data on
individual trees forming a forest and a polygon that defines the border of
the forest. The objects are used as an input for simulating forest
inventories.
}
\details{
The following columns are needed as a minimum requirement in
  \code{data}:
  \itemize{
  \item \code{id_tree} - integer values used to
  identify individual trees
  \item \code{id_stem} - integer values to identify
  individual stems. A tree can fork into multiple stems and stems are treated
  as individual record under certain circumstances (exact definitions depend
  on the particular surveys at which trees were measured). If a tree has two
  stems, for example, \code{id_tree} occurs twice for that tree, while
  \code{id_stem} is unique for each stem.
  \item \code{x_tree} The x-coordinates of the trees, defined at the stem
  center at a height of 1.3 meter above ground level. The data type is
  numeric (double) and the units are usually in meter.
  \item \code{y_tree} The y-coordinates of the trees, defined at the stem
  center at a height of 1.3 meter above ground level. The data type is
  numeric (double) and the units are usually in meter.
  \item \code{is_dead} A logical indicating whether the stem is dead
  (\code{TRUE}) or alive (\code{FALSE}).
  \item \code{dbh} The diameter of the stem, measured at a height of
  1.3 meter above ground in cm (diameter at breast height, \code{numeric}).
  }

  During object initialization, three variables required for simulations will
  be added:
  \itemize{
  \item \code{ba} The cross-section of the stems at a height of 1.3 meter
  above ground level calculated from the provided \code{dbh} values assuming
  a circular stem.
  \item \code{n} The number of stems; initialized with one for each entry.
  This is simply a helper variable for simualtions and estimation.
  \item \code{f_edge} Edge factor, used for edge correction of trees close
  to the forest border. Since such trees have a lower probability of being
  included into the sample due to the restriction of the sampling to the
  study area defined by \code{boundary}, edge correction must be performed
  to achieve unbiase estimation. The variable is initialized with one and
  will be modified during the simulation according to appropriate edge
  correction methods chosen by the user. After edge correction, some trees
  are counted twice to compensate for the lower inlcusion probabilities of
  border trees.
  }

  This minimum set of attributes is required for the estimation of stem
  counts, basal area (a measure for site occupation) and diameter
  distributions. Users may add additional variables like species,
  tree volume, aboveground biomass, etc. It will be possible to include such
  addtional variables in the simulation.
}
