% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{tree_sample}
\alias{tree_sample}
\title{The function \code{tree_sample} creates tree sample objects. Such objects
assign individual tree information from the population to the sample
locations following a particular response design. The function is only used
internally by \code{\link{extract_data}}.}
\usage{
tree_sample(data, r_design, r_design_parm)
}
\arguments{
\item{data}{A \code{\link[data.table]{data.table}} object holding individual
tree data together with identifiers that indicate the sample and the point
location at which trees were selected.}

\item{r_design}{A \code{character} string indicating the type of the response
design. One of the following: "fixed_area", "k_tree", "angle_count".}

\item{r_design_parm}{A response design specific parameter of type
\code{numeric}. For "fixed_area" the plot radius in meter; for "k_tree" the
number of trees closest to the sample location; for "angle_count" the basal
area factor.}
}
\value{
\enumerate{
  \item \code{$data} A \code{\link[data.table]{data.table}}
  object as described above.
  \item \code{$r_design}
  \item \code{$r_design_parm}
  }
}
\description{
The function \code{tree_sample} creates tree sample objects. Such objects
assign individual tree information from the population to the sample
locations following a particular response design. The function is only used
internally by \code{\link{extract_data}}.
}
\details{
The \code{\link[data.table]{data.table}} object (\code{data}) should
  have the following columns:
  \itemize{
  \item \code{id_sample}
  \item \code{id_point}
  \item \code{id_tree}
  \item \code{id_stem}
  \item \code{x_tree}
  \item \code{y_tree}
  \item \code{is_dead}
  \item \code{dbh}
  \item \code{ba}
  \item \code{n}
  \item \code{f_edge}
  \item \code{ef}
  }

  Some response designs (currently only \code{\link{k_tree}}), may have
  additional alternative expansion factors as optional approximations, where
  unbiased estimators are missing or difficult to derive. As the column names
  for these alternatives \code{ef_alt1} and \code{ef_alt2} are used.
}
