% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_design.R
\name{xy_sample}
\alias{xy_sample}
\title{Sample locations in a study area of rectangular shape}
\usage{
xy_sample(x_range, y_range, n)
}
\arguments{
\item{x_range}{A numeric vector holding minimum and maximum coordinate in
x-direction.}

\item{y_range}{A numeric vector holding minimum and maximum coordinate in
y-direction.}

\item{n}{Sample size, i.e. number of sample locations.}
}
\value{
A \code{\link{matrix}} object with \code{n} rows and
  x-coordinates in the first column and y-coordinates in the second.
}
\description{
Sample locations in a study area of rectangular shape
}
