% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_design.R
\name{xy_sample_regular}
\alias{xy_sample_regular}
\title{Place a regular sampling grid over a polygon of arbitrary shape}
\usage{
xy_sample_regular(sf_poly, n, M = 1, cell_size = NULL, random_rot = TRUE)
}
\arguments{
\item{sf_poly}{An object of class \code{\link[sf]{sf}} and geometry type
POLYGON from the \code{sf} package.}

\item{n}{Sample size}

\item{M}{Number of independent samples of size \code{n}}

\item{cell_size}{If missing, a cell size is derived from the sample size
\code{n}; otherwise the cell size is used to calculate the grid spacing
(assuming a square grid).}

\item{random_rot}{Logical, if TRUE the grid is rotated randomly}
}
\value{
A \code{\link[data.table]{data.table}} object with approximately
  \code{M} times \code{n} rows holding an identifier and xy-coordinates.
}
\description{
Place a regular sampling grid over a polygon of arbitrary shape
}
\details{
Due to the systematic nature of the generated samples, sample size
  is not constant but will on average reach the expected sample size
  \code{n}.

  Grids are generated with a random starting point choosen from an area in
  the lower left corner of the polygon's bounding box. The size of this area
  corresponds to the cell size of the grid as set by \code{cell_size} or as
  inferred from \code{n} if \code{cell_size} is \code{NULL}.

  Sample grids are initially generated for the entire bounding box that
  encompasses \code{sp_poly}, and are refined to the actual outline of the
  study region in a second step. If \code{M} becomes large (say > 10000), the
  computation of the \code{M} samples will take while.
}
