% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edge_corr_wt}
\alias{edge_corr_wt}
\title{Performs the \emph{walkthrough} edge-correction method for a set of sample
trees and a given boundary}
\usage{
edge_corr_wt(tree_pop, tree_sample, sample_loc)
}
\arguments{
\item{tree_pop}{A \code{\link{tree_pop}} object.}

\item{tree_sample}{A \code{\link{tree_sample}} object as coming from the
\code{\link{extract_data}} function.}

\item{sample_loc}{A \code{\link{sample_loc}} object as created by the
\code{\link{xy_sample}} function or provided by the user.}
}
\value{
The \code{tree_sample} object with modified \code{f_edge} column
  indicating which trees are counted twice.
}
\description{
Performs the \emph{walkthrough} edge-correction method for a set of sample
trees and a given boundary
}
\references{
Ducey, M.J., Gove, J.H., Valentine, H.T., 2004. A walkthrough
  solution to the boundary overlap problem. Forest Science 50, 427–435.
}
