% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_area}
\alias{extract_area}
\title{Helper function to extract the area slots from
\code{\link[sp]{SpatialPolygons}} objects}
\usage{
extract_area(sp_poly)
}
\arguments{
\item{sp_poly}{An object of class \code{\link[sp]{SpatialPolygons}}}
}
\value{
A numeric vector with the areas of all polygons in the object. The
  areas for polygons that represent holes are stored as negative values.
}
\description{
Helper function to extract the area slots from
\code{\link[sp]{SpatialPolygons}} objects
}
\examples{
library(sp);
library(maptools);
data("state.vbm");
a <- extract_area(state.vbm);
}
