% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_designs.R
\name{fixed_area}
\alias{fixed_area}
\title{Select trees at sample locations using fixed-area plots}
\usage{
fixed_area(tree_pop, sample_loc, r, k = NULL)
}
\arguments{
\item{tree_pop}{A tree population object (\code{\link{tree_pop}})}

\item{sample_loc}{A sample location object (\code{\link{sample_loc}}) as
provided by the \code{\link{xy_sample}} function.}

\item{r}{The radius of the fixed-area sample plot in the same units as the
coorinates in \code{tree_loc} and \code{sample_loc}.}

\item{k}{Number of neighbors to search for within \code{r}. See details.}
}
\value{
An object of class \code{\link{response}} with indices of trees in
  \code{tree_pop} that are selected at the individual sample locations.
}
\description{
Select trees at sample locations using fixed-area plots
}
\details{
The RANN package is used to find all trees within the specified
  radius. The RANN package is a wrapper for the Approximate Near Neighbor
  (ANN) C++ library allowing for fast nearest-neighbor searches using
  kd-trees. Here, radius search is used, meaning that the \code{k}-nearest
  neighbors within \code{r} are searched for. To find all trees within
  \code{r}, \code{k} should be large enough (max the number of rows in
  \code{tree_loc}). If not specified by the user, \code{k} is estimated by
  multiplying the average tree density with plot area. Tree density is
  inferred using the \code{density.ppp} function from the spatstat package.
  Note that the estimation costs some extra computation time and it is thus
  recommended to provide a large enough \code{k} when the function is used in
  simulations.
}
