% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_designs.R
\name{k_tree}
\alias{k_tree}
\title{The \code{k} trees closest to a sample location are selected into the sample}
\usage{
k_tree(tree_pop, sample_loc, k = 6)
}
\arguments{
\item{tree_pop}{A tree population object (\code{\link{tree_pop}})}

\item{sample_loc}{A sample location object (\code{\link{sample_loc}}) as
provided by the \code{\link{xy_sample}} function.}

\item{k}{Number of nearets neighbors.}
}
\value{
An object of class \code{\link{response}} with indices of trees in
  \code{tree_pop} that are selected at the individual sample locations.
}
\description{
The \code{k} trees closest to a sample location are selected into the sample
}
\details{
The RANN package is used to find the \code{k} trees clostest to the
  respective sample locations in \code{sample_loc}. The RANN package is a
  wrapper for the Approximate Near Neighbor (ANN) C++ library allowing for
  fast nearest-neighbor searches using kd-trees.
}
