% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kalimantan_peat}
\alias{kalimantan_peat}
\title{Full census of peat swamp forest in Kalimantan on Borneo island}
\format{A data.table with 3523 rows and 8 columns: \describe{
  \item{stem_id}{Stem number - one tree can have mulitple stems}
  \item{tree_id}{Tree number - repetitions possible due to multiple stems}
  \item{dbh}{Diameter of the stem measured at breast height (1.3m) in cm}
  \item{tree_height}{Height of the trees in m} \item{crownwidth}{Width of the
  tree cronwns in m} \item{dead}{Was the tree alive or dead?}
  \item{xrel}{Relative position of the stem center in x-direction}
  \item{yrel}{Relative position of the stem center in y-direction}
  \item{ba}{Cross-sectional area of the stem at a height of 1.3m in square
  meter}
  \item{n}{Stem count - helper variable for stem number estimation, unitless}}}
\source{
\url{https://www.uni-goettingen.de/en/67094.html}
}
\usage{
kalimantan_peat
}
\description{
A dataset containing tree stem locations and other attributes of 3523 trees
measured on 120m by 120m large full census plot in a peat swamp forest on
Borneo island.
}
\keyword{datasets}
